complete sign in sign up

+++ src/app/components/admin/dashboard/dashboard.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
-import { IProduct } from 'src/app/models/products';
+import { lastValueFrom } from 'rxjs';
+import { Data, IProduct } from 'src/app/models/products';
 import { ProductServiceService } from 'src/app/services/product-service.service';
 @Component({
   selector: 'app-dashboard',
@@ -7,24 +8,39 @@ import { ProductServiceService } from 'src/app/services/product-service.service'
   styleUrls: ['./dashboard.component.css']
 })
 export class DashboardComponent {
-  products:any =[]
+  products:IProduct[] = []
   constructor(private productService:ProductServiceService){
     this.loadProducts()
   }
-  loadProducts(){
-    this.productService.getAll().subscribe({
-      next:(data)=>this.products = data,
-      error:(err)=>console.log(err)
-    })
-  }
-  handleDelete(id:number){
-    this.productService.deleteProduct(id).subscribe({
-      next:(data)=>{
-        this.loadProducts()
-        alert("delete product successfully")
-      },
-      error : (err)=>console.log(err)
+  async loadProducts(){
 
-    })
+    const data: Data = await lastValueFrom(this.productService.getAll())
+    this.products = data.allProduct?.docs || [];
+    
+    // this.productService.getAll().subscribe({
+    //   next:(data: {
+    //       message: string, 
+    //       allProduct: { 
+    //         docs: IProduct[]
+    //       }
+    //     })=>{
+    //       this.products = data.allProduct?.docs || [] ;
+    //       console.log(this.products)
+    //     },
+    //   error:(err)=>console.log(err)
+    // })
+  }
+  handleDelete(id:number|undefined){
+    if(id){
+      this.productService.deleteProduct(id).subscribe({
+        next:(data)=>{
+          this.loadProducts()
+          alert("delete product successfully")
+        },
+        error : (err)=>console.log(err)
+  
+      })
+    }
+   
   }
 }
